Another perf PR inspired by #1078.

Running with [JRuby's --profile option](https://github.com/jruby/jruby/wiki/Profiling-jruby) revealed that we were spending a really huge amount of time in the various `Dimension` methods which were being meta-programmed up. It was up to 5x faster by my benchmarking to just do things directly, so I've unrolled the meta in favor of speed.