@jasonrclark 2015-03-24
FWIW, when looking at a merge with both this and #1079, perf on the colors page in the manual was back up to `pre2` levels for me (anecdotally).

I'm interested to continue peeling back a bit more on the perf onion, though, to see if there's any other easy, low-hanging fruit that can get large apps like the manual running smoothly rather than just acceptably.