@PragTob 2015-03-24
Thanks a lot for looking at this and fixing this :)

I'm curios though why there is such a big difference. Sure those methods are called a lot and they do more lookups to dynamically get the instance variable... I wonder if we could use the `define_method` version with an interpolated string to get rid off this "extra work" we did, because then theoretically that defined method should be the same as the written out versions.

Nothing to change here, just something for my personal learning and interest