@jasonrclark 2015-03-24
I was as surprised as you are the impact of it. When I do the benchmarking I'll do a bit of comparison, but I found that actually each bit of the meta-work had a significant burden by itself over just doing it directly. Interesting stats to come in the near future, and thanks for asking about it! :)