@jasonrclark 2015-03-25
Also rebased off master to provide [`benchmark/dimension.rb`](https://github.com/shoes/shoes4/blob/master/benchmark/dimension.rb) before and after

**Before**
```
✔ ♥♥♥♥ [jruby] (master)*
~/source/shoes4:bin/shoes benchmark/dimension.rb
       user     system      total        real
1_000_000x margin_start  3.040000   0.050000   3.090000 (  2.727000)
1_000_000x element_extent  6.810000   0.030000   6.840000 (  6.509000)
```

**After**
```
✔ ♥♥♥♥ [jruby] (less-meta-dimension)*
~/source/shoes4:bin/shoes benchmark/dimension.rb
       user     system      total        real
1_000_000x margin_start  0.540000   0.010000   0.550000 (  0.473000)
1_000_000x element_extent  1.970000   0.020000   1.990000 (  1.885000)
```

So that's a pretty big drop. You might wonder whether it matters, but here's a snippet of profiling from running the manual with some call counts of interest:

```
 %total   %self       total        self    children                 calls  name
---------------------------------------------------------------------------------------------------------
                       1.51        0.02        1.50       3627811/3924939  Shoes::Dimension#calculate_relative
                      10.75        0.03       10.72         19184/3924939  BasicObject#__send__
                      44.89        0.21       44.68        277944/3924939  Shoes::Dimension#element_end
  7727%     34%       57.15        0.26       56.90               3924939  Shoes::Dimension#element_extent
                      46.71        0.59       46.12       7773908/7900855  Shoes::Dimension#extent
                      39.83        0.07       39.75           38728/42527  Shoes::ParentDimension#extent
                      16.49        7.14        9.35       3887697/4421200  Shoes::Dimension#margin_start
                      14.57        6.47        8.09       3887697/3969633  Shoes::Dimension#margin_end
                       0.24        0.24        0.00       3887697/4657600  Kernel.nil?
                       0.00        0.00        0.00          37242/184482  NilClass#nil?
```

And yes, you read that right--`Shoes::Dimension#extent` was called 8 million times! Obviously we should do some hard looking at our layout paths to see if we can reduce that algorithmically, but regardless what we do I think that `Dimension` is bound to always be in our hot path.

@PragTob, to your question about where the time is going, I've put up a [sample gist here](https://gist.github.com/jasonrclark/1f71e5162ba8fbc77af7) that takes one of the more meta methods (`margin_start) and step by step deconstructs it to direct access to see the impact of each bit of "meta" we're using.

* Fully meta-programmed (`full_meta`)
* Extracting the ivar name (`hoist_ivar_name`)
* Directly accessing the ivar instead of `instance_variable_get` (`direct_ivar`)
* Memoizing the relative flagging rather than checking repeatedly (`memoize_relative_flag`)
* `def` instead of `define_method`, with everything done directly (`full_direct`)

This uses the most excellent [benchmark-ips](https://github.com/evanphx/benchmark-ips) which I find much easier for these type of comparisons than plain `benchmark`. The results:

```
Calculating -------------------------------------
           full_meta    16.473k i/100ms
     hoist_ivar_name    17.984k i/100ms
         direct_ivar    28.890k i/100ms
memoize_relative_flag
                        33.238k i/100ms
         full_direct    36.262k i/100ms
-------------------------------------------------
            full_meta    325.594k (± 5.8%) i/s -      1.631M
      hoist_ivar_name    423.482k (± 3.3%) i/s -      2.122M
          direct_ivar    798.759k (± 3.2%) i/s -      3.987M
memoize_relative_flag      1.208M (± 3.0%) i/s -      6.049M
          full_direct      1.523M (± 3.9%) i/s -      7.615M

Comparison:
          full_direct:  1522752.7 i/s
memoize_relative_flag:  1207850.6 i/s - 1.26x slower
          direct_ivar:   798758.6 i/s - 1.91x slower
      hoist_ivar_name:   423482.2 i/s - 3.60x slower
            full_meta:   325593.6 i/s - 4.68x slower

```

So that's the basis for my proposed changes here. I'm hoping that we can cut things down in a big way algorithmically as well at some point, but this at least gives us the fastest basis (or close at least) in `Dimension` that I think we can hope for.

Shoes on!