@PragTob 2015-03-25
Hi there! Awesome break down work! :D You should really get that into a blog post, if you won't, then I will! :)

That's some great performance work right there and I agree... I'm still looking forward very much to algorithmic changes on this e.g. don't do full contents_alignment all the time, that's one of my favorite tickets waiting right there.

Looking at this I remembered a project I was once one and was wondering why people used eval and def instead of the meta programming `define_method` - and stuff like this is probably why. I forked your gist [and added one variant](https://gist.github.com/PragTob/eb43aaa99e4b86b99d4e) where I do an eval and `def` on a string. it performs as well as `full_direct`. We could interpolate that string and with that keep the meta programming around. We don't have to, Metaprogramming is often evil and with that we lose code highlighting as well - just if the duplication bugs us again. Your call :)

Also please add `benachmark/ips` I love it myself - no idea why I didn't use it with the benchmarks before. Maybe I didn't know it back then.

Really, AMAZING work here :star: 

Shoes on!
Tobi