@jasonrclark 2015-03-27
Thanks @PragTob, and interesting other stuff you turned up on JRuby there! Have been meaning to blog more (like everyone I think :stuck_out_tongue:), and this would be a good thing to write up.

Merging so I can run some things off master with both perf branches landed.